Implemented tests for api.py and fixed bugs in api.py.

+++ src/melvonaut/api.py
@@ -1,6 +1,7 @@
 import importlib.metadata
 import psutil
 from aiohttp import web
+from io import StringIO, BytesIO
 import melvonaut.settings as settings
 import melvonaut.persistent_settings as p_settings
 from shared import constants as con
@@ -8,6 +9,8 @@ from loguru import logger
 import asyncio
 import pathlib
 import shared.models as models
+import datetime
+from melvonaut.settings import setup_file_logging
 
 
 async def health(request: web.Request):
@@ -53,15 +56,18 @@ async def get_restart_melvin(request: web.Request):
 
 
 async def get_shutdown_melvin(request: web.Request):
-    loop = asyncio.get_running_loop()
-    loop.stop()
-    pending_tasks = asyncio.all_tasks()
-    for task in pending_tasks:
-        task.cancel()
     try:
         return web.Response(status=200, text="OK")
     finally:
-        exit()
+        if settings.DO_ACTUALLY_EXIT:
+            loop = asyncio.get_running_loop()
+            loop.stop()
+            pending_tasks = asyncio.all_tasks()
+            for task in pending_tasks:
+                task.cancel()
+            exit()
+        else:
+            logger.warning("Requested shutdown, but not actually exiting")
 
 
 async def post_execute_command(request: web.Request):
@@ -101,16 +107,23 @@ async def get_melvin_version(request: web.Request):
 async def get_list_log_files(request: web.Request):
     logger.debug("Listing log files")
     log_files = []
-    folder = pathlib.Path(con.MEL_LOG_LOCATION)
-    for file in folder.iterdir():
-        log_files.append(file.name)
+    folder = pathlib.Path(con.MEL_LOG_PATH)
+    try:
+        for file in folder.iterdir():
+            if not file.is_file():
+                continue
+            if not file.name.endswith(".log"):
+                continue
+            log_files.append(file.name)
+    except Exception as e:
+        return web.json_response({"error": str(e)}, status=500)
     return web.json_response({"log_files": log_files}, status=200)
 
 
 async def post_download_log(request: web.Request):
     data = await request.json()
     logger.debug(f"Downloading log: {data}")
-    log_file = pathlib.Path(con.MEL_LOG_LOCATION) / data.get("file")
+    log_file = pathlib.Path(con.MEL_LOG_PATH) / data.get("file")
     if log_file.exists():
         return web.FileResponse(log_file, status=200)
     else:
@@ -120,12 +133,14 @@ async def post_download_log(request: web.Request):
 async def post_download_log_and_clear(request: web.Request):
     data = await request.json()
     logger.debug(f"Downloading log and clearing: {data}")
-    log_file = pathlib.Path(con.MEL_LOG_LOCATION) / data.get("file")
+    log_file = pathlib.Path(con.MEL_LOG_PATH) / data.get("file")
     if log_file.exists():
+        log_file_content = StringIO(log_file.read_text())
         try:
-            return web.FileResponse(log_file, status=200)
+            return web.Response(body=log_file_content, status=200)
         finally:
             log_file.unlink()
+            setup_file_logging()
     else:
         return web.Response(status=404, text="File not found")
 
@@ -133,22 +148,36 @@ async def post_download_log_and_clear(request: web.Request):
 async def post_clear_log(request: web.Request):
     data = await request.json()
     logger.debug(f"Clearing log: {data}")
-    log_file = pathlib.Path(con.MEL_LOG_LOCATION) / data.get("file")
+    log_file = pathlib.Path(con.MEL_LOG_PATH) / data.get("file")
     if log_file.exists():
+        if log_file.is_dir():
+            return web.Response(status=400, text=f"{log_file} is a directory")
+        if not log_file.name.endswith(".log"):
+            return web.Response(status=400, text=f"{log_file} is not a log file")
         log_file.unlink()
-        return web.Response(status=200, text="OK")
+        setup_file_logging()
+        return web.Response(status=200, text=f"{log_file} cleared")
     else:
-        return web.Response(status=404, text="File not found")
+        return web.Response(status=404, text=f"{log_file} not found")
 
 
 async def get_clear_all_logs(request: web.Request):
-    logger.debug("Clearing all log files")
-    log_files = []
-    folder = pathlib.Path(con.MEL_LOG_LOCATION)
-    for file in folder.iterdir():
-        log_files.append(file.name)
-        file.unlink()
-    return web.json_response({"Cleared_files": log_files}, status=200)
+    try:
+        logger.debug("Clearing all log files")
+        log_files = []
+        folder = pathlib.Path(con.MEL_LOG_PATH)
+        for file in folder.iterdir():
+            if file.is_dir():
+                continue
+            if not file.name.endswith(".log"):
+                continue
+            log_files.append(file.name)
+            file.unlink()
+        setup_file_logging()
+        return web.json_response({"Cleared_files": log_files}, status=200)
+    except Exception as e:
+        logger.exception(e)
+        return web.json_response({"error": str(e)}, status=500)
 
 
 # Download telemetry
@@ -165,8 +194,9 @@ async def get_download_telemetry_and_clear(request: web.Request):
     logger.debug("Downloading telemetry and clearing")
     telemetry_file = pathlib.Path(con.TELEMETRY_LOCATION_CSV)
     if telemetry_file.exists():
+        telemetry_file_content = StringIO(telemetry_file.read_text())
         try:
-            return web.FileResponse(telemetry_file, status=200)
+            return web.Response(body=telemetry_file_content, status=200)
         finally:
             telemetry_file.unlink()
     else:
@@ -197,8 +227,9 @@ async def get_download_events_and_clear(request: web.Request):
     logger.debug("Downloading events and clearing")
     events_file = pathlib.Path(con.EVENT_LOCATION_CSV)
     if events_file.exists():
+        events_file_content = StringIO(events_file.read_text())
         try:
-            return web.FileResponse(events_file, status=200)
+            return web.Response(body=events_file_content, status=200)
         finally:
             events_file.unlink()
     else:
@@ -218,7 +249,9 @@ async def get_clear_events(request: web.Request):
 async def get_list_images(request: web.Request):
     logger.debug("Listing images")
     folder = pathlib.Path(con.IMAGE_PATH_BASE)
-    images = [str(file) for file in folder.rglob("*.png") if file.is_file()]
+    if not folder.exists():
+        return web.Response(status=404, text=f"Folder not found: {folder}")
+    images = [str(file.name) for file in folder.rglob("*.png") if file.is_file()]
     return web.json_response({"images": images}, status=200)
 
 
@@ -237,8 +270,9 @@ async def post_download_image_and_clear(request: web.Request):
     logger.debug(f"Downloading image and clearing: {data}")
     image_file = pathlib.Path(con.IMAGE_PATH_BASE) / data.get("file")
     if image_file.exists():
+        image_file_content = BytesIO(image_file.read_bytes())
         try:
-            return web.FileResponse(image_file, status=200)
+            return web.Response(body=image_file_content, status=200)
         finally:
             image_file.unlink()
     else:
@@ -257,12 +291,20 @@ async def get_clear_all_images(request: web.Request):
 async def post_set_melvin_task(request: web.Request):
     data = await request.json()
     logger.debug(f"Setting melvin task: {data}")
-    task = data.get("task")
+    task = data.get("task", None)
+    if not task:
+        logger.warning("Missing field task")
+        return web.Response(status=400, text="Missing field task")
     try:
         melvin_task = models.MELVINTask(task)
     except ValueError:
+        logger.warning("Invalid task")
         return web.Response(status=400, text="Invalid task")
+    except Exception as e:
+        logger.warning(f"Error setting melvin task: {e}")
+        return web.Response(status=500, text=str(e))
     await p_settings.set_settings({"CURRENT_MELVIN_TASK": melvin_task})
+    return web.Response(status=200, text="OK")
 
 
 async def get_reset_settings(request: web.Request):
@@ -287,25 +329,34 @@ async def post_clear_setting(request: web.Request):
     return web.Response(status=200, text="OK")
 
 
-async def get_setting(request: web.Request):
+async def post_get_setting(request: web.Request):
     logger.debug("Getting settings")
     data = await request.json()
     logger.debug(f"Requested settings: {data}")
     response_settings = {}
     loaded_settings = await p_settings.load_settings()
     for key, value in data.items():
-        response_settings[key] = loaded_settings.get(
-            key, getattr(settings, key.upper())
-        )
-    return web.json_response(get_reset_settings, status=200)
+        try:
+            response_settings[key] = loaded_settings.get(
+                key, getattr(settings, key.upper())
+            )
+        except AttributeError:
+            return web.Response(status=404, text=f"Setting not found: {key}")
+    return web.json_response(response_settings, status=200)
 
 
 async def get_all_settings(request: web.Request):
     logger.debug("Getting all settings")
-    attrs = dir(settings)
+    attrs = [key for key in dir(settings) if not key.startswith("_") and key.isupper()]
     all_settings = {}
     for attr in attrs:
-        all_settings[attr] = getattr(settings, attr)
+        value = getattr(settings, attr)
+        if type(value) is float or type(value) is int:
+            all_settings[attr] = value
+        elif type(value) is datetime.datetime:
+            all_settings[attr] = value.isoformat()
+        else:
+            all_settings[attr] = str(value)
     additional_settings = await p_settings.load_settings()
     for key, value in additional_settings.items():
         all_settings[key] = value
@@ -313,23 +364,24 @@ async def get_all_settings(request: web.Request):
 
 
 def setup_routes(app) -> None:
-    app.router.add_post("/api/download_logs", post_download_log)
+    app.router.add_post("/api/download_log", post_download_log)
     app.router.add_get("/api/download_telemetry", get_download_telemetry)
     app.router.add_get("/api/download_events", get_download_events)
-    app.router.add_get("/api/download_images", post_download_image)
+    app.router.add_post("/api/download_image", post_download_image)
     app.router.add_post("/api/set_melvin_task", post_set_melvin_task)
     app.router.add_get("/api/reset_settings", get_reset_settings)
     app.router.add_post("/api/set_setting", post_set_setting)
     app.router.add_post("/api/clear_setting", post_clear_setting)
-    app.router.add_post("/api/get_setting", get_setting)
+    app.router.add_post("/api/clear_log", post_clear_log)
+    app.router.add_post("/api/get_setting", post_get_setting)
     app.router.add_get("/api/get_all_settings", get_all_settings)
-    app.router.add_get("/api/download_logs_and_clear", post_download_log_and_clear)
+    app.router.add_post("/api/download_log_and_clear", post_download_log_and_clear)
     app.router.add_get(
         "/api/download_telemetry_and_clear", get_download_telemetry_and_clear
     )
     app.router.add_get("/api/download_events_and_clear", get_download_events_and_clear)
-    app.router.add_get("/api/download_images_and_clear", post_download_image_and_clear)
-    app.router.add_get("/api/clear_logs", get_clear_all_logs)
+    app.router.add_post("/api/download_image_and_clear", post_download_image_and_clear)
+    app.router.add_get("/api/clear_all_logs", get_clear_all_logs)
     app.router.add_get("/api/clear_telemetry", get_clear_telemetry)
     app.router.add_get("/api/clear_events", get_clear_events)
     app.router.add_get("/api/clear_images", get_clear_all_images)
@@ -342,6 +394,7 @@ def setup_routes(app) -> None:
     app.router.add_post("/api/execute_command", post_execute_command)
     app.router.add_get("/api/get_melvin_version", get_melvin_version)
     app.router.add_get("/api/list_log_files", get_list_log_files)
+    app.router.add_get("/api/list_images", get_list_images)
 
 
 async def run_api() -> None:
